+++ source/dotnet/wholesale-api/Batches/Batches.Application/Model/BatchDtoMapper.cs
@@ -12,22 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Application.Processes.Model;
+using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
+using Energinet.DataHub.Wholesale.Batches.Infrastructure.GridAreaAggregate;
+using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
 using Energinet.DataHub.Wholesale.Contracts;
-using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
-using Energinet.DataHub.Wholesale.Domain.GridAreaAggregate;
 
-namespace Energinet.DataHub.Wholesale.Application.Batches.Model;
+namespace Energinet.DataHub.Wholesale.Batches.Application.Model;
 
 public class BatchDtoMapper : IBatchDtoMapper
 {
-    private readonly IProcessTypeMapper _processTypeMapper;
-
-    public BatchDtoMapper(IProcessTypeMapper processTypeMapper)
-    {
-        _processTypeMapper = processTypeMapper;
-    }
-
     public BatchDto Map(Batch batch)
     {
         return new BatchDto(
@@ -42,7 +35,7 @@ public class BatchDtoMapper : IBatchDtoMapper
             MapState(batch.ExecutionState),
             batch.AreSettlementReportsCreated,
             MapGridAreaCodes(batch.GridAreaCodes),
-            _processTypeMapper.MapFrom(batch.ProcessType));
+            ProcessTypeMapper.MapFrom(batch.ProcessType));
     }
 
     private static BatchState MapState(BatchExecutionState state)
