+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/BatchAggregate/Batch.cs
@@ -13,8 +13,6 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Batches.Infrastructure.GridAreaAggregate;
-using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
-using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 using NodaTime;
 
 namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
