+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/BatchAggregate/CalculationId.cs
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Domain.BatchAggregate;
+namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
 
 public sealed record CalculationId(long Id);
