+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/BatchAggregate/IBatchRepository.cs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
 using Energinet.DataHub.Wholesale.Batches.Infrastructure.GridAreaAggregate;
+using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 using NodaTime;
 
-namespace Energinet.DataHub.Wholesale.Domain.BatchAggregate;
+namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
 
 public interface IBatchRepository
 {
