+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/CalculationDomainService/CalculationState.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Domain.CalculationDomainService;
+namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.CalculationDomainService;
 
 public enum CalculationState
 {
