+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/CalculationDomainService/ICalculationEngineClient.cs
@@ -13,9 +13,8 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchAggregate;
-using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 
-namespace Energinet.DataHub.Wholesale.Domain.CalculationDomainService;
+namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.CalculationDomainService;
 
 public interface ICalculationEngineClient
 {
