+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/Persistence/IUnitOfWork.cs
@@ -12,23 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Contracts;
+namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.Persistence;
 
-namespace Energinet.DataHub.Wholesale.Application.Batches.Model;
-
-/// <summary>
-/// An immutable batch.
-/// </summary>
-public sealed record BatchDto(
-    long? RunId,
-    Guid BatchId,
-    DateTimeOffset PeriodStart,
-    DateTimeOffset PeriodEnd,
-    string Resolution,
-    string Unit,
-    DateTimeOffset? ExecutionTimeStart,
-    DateTimeOffset? ExecutionTimeEnd,
-    BatchState ExecutionState,
-    bool AreSettlementReportsCreated,
-    string[] GridAreaCodes,
-    ProcessType ProcessType);
+public interface IUnitOfWork
+{
+    Task CommitAsync();
+}
