+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/Persistence/UnitOfWork.cs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Batches.Application;
-
 namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.Persistence;
 
 public class UnitOfWork : IUnitOfWork
