+++ source/dotnet/wholesale-api/Batches/Batches.Interfaces/IBatchApplicationService.cs
@@ -12,18 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Application.Batches.Model;
+using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
 using Energinet.DataHub.Wholesale.Contracts;
-using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 
-namespace Energinet.DataHub.Wholesale.Application.Batches;
+namespace Energinet.DataHub.Wholesale.Batches.Interfaces;
 
 public interface IBatchApplicationService
 {
-    /// <summary>
-    /// Create and start all processes of batches with state <see cref="BatchExecutionState.Submitted"/>.
-    /// </summary>
-    /// <param name="batchId"></param>
     Task StartCalculationAsync(Guid batchId);
 
     Task UpdateExecutionStateAsync();
