+++ source/dotnet/wholesale-api/Batches/Batches.Interfaces/Models/BatchRequestDto.cs
@@ -12,15 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Contracts;
+using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
 
-namespace Energinet.DataHub.Wholesale.Application.Batches;
+namespace Energinet.DataHub.Wholesale.Contracts;
 
 /// <summary>
-/// An immutable command to create a batch.
+/// An immutable request to create a batch.
 /// </summary>
-public sealed record CreateBatchCommand(
-    ProcessType ProcessType,
-    IEnumerable<string> GridAreaCodes,
-    DateTimeOffset StartDate,
-    DateTimeOffset EndDate);
+public sealed record BatchRequestDto(ProcessType ProcessType, IEnumerable<string> GridAreaCodes, DateTimeOffset StartDate, DateTimeOffset EndDate);
