+++ source/dotnet/wholesale-api/Batches/Batches.Interfaces/Models/BatchSearchDto.cs
