+++ source/dotnet/wholesale-api/Batches/Batches.Interfaces/Models/BatchSearchDtoV2.cs
