+++ source/dotnet/wholesale-api/Batches/Batches.Interfaces/Models/BatchState.cs
