+++ source/dotnet/wholesale-api/Batches/Batches.Interfaces/Models/ProcessType.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
+namespace Energinet.DataHub.Wholesale.Contracts;
 
 /// <summary>
 /// Defines the wholesale process type
