+++ source/dotnet/wholesale-api/WebApi/V3/ProcessStepResult/IProcessStepResultFactory.cs
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Wholesale.Application.Batches.Model;
+
+using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
 
 namespace Energinet.DataHub.Wholesale.WebApi.V3.ProcessStepResult;
